fix: implement cross-fork is_parent_strong()
